refactor: relocate page contracts to Core and enforce layering
Move PageContract and PageMetadata out of the Http orchestration layer into
Core, where the lower layers that need them (Core discovery, Layouts menu)
already live. Http previously sat below the feature domains in the documented
model yet consumed them, so Core and Layouts were depending upward into it.
Retype PageMetadata's factories to the PageContract interface to drop the last
reference to the abstract Http\Page.

Correct the ArchTest header to describe the real layering (Http/Pages are
orchestration layers above the domains) and add guards that encode it: Core and
the feature domains may not depend upward on Http/Pages/Console/Facades, and
Attributes stays free of the domains. Document why co-located capability
interfaces are exempt from the Contracts convention.
