+++ src/Http/Page.php
@@ -8,7 +8,9 @@ use BadMethodCallException;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 use Inertia\Response;
+use Lattice\Lattice\Core\Contracts\PageContract;
 use Lattice\Lattice\Core\Enums\PageLayout;
+use Lattice\Lattice\Core\PageMetadata;
 use Lattice\Lattice\Core\PageSchema;
 use Lattice\Lattice\Facades\Lattice;
 use UnexpectedValueException;
