+++ tests/ArchTest.php
@@ -2,12 +2,19 @@
 declare(strict_types=1);
 
 /*
- * Feature-domain isolation.
+ * Layering.
  *
- * Each feature domain builds on the shared layers (Core, Http, Attributes,
- * Support) but stays independent of its sibling domains. The only intentional
- * couplings are tables -> actions (row and bulk actions) and actions -> forms
- * (action forms); every other cross-domain edge is forbidden below.
+ * Bottom: the shared base — Core, Attributes, and the Support utilities — which
+ * the rest of the package builds on and which never depend back on a feature
+ * domain or an orchestration layer.
+ *
+ * Middle: the five feature domains. Each stays independent of its siblings; the
+ * only intentional cross-domain couplings are tables -> actions (row and bulk
+ * actions) and actions -> forms (action forms).
+ *
+ * Top: the orchestration and tooling layers — Http and Pages (which render and
+ * route pages by consuming the feature domains), Console, and Facades. Nothing
+ * below may depend upward on them.
  */
 
 arch('forms depend on no other feature domain')
@@ -63,10 +70,58 @@ arch('core does not depend on the feature domains')
         'Lattice\Lattice\Layouts',
     ]);
 
+arch('core does not depend upward on the orchestration or tooling layers')
+    ->expect('Lattice\Lattice\Core')
+    ->not->toUse([
+        'Lattice\Lattice\Http',
+        'Lattice\Lattice\Pages',
+        'Lattice\Lattice\Console',
+        'Lattice\Lattice\Facades',
+    ]);
+
+arch('feature domains never depend upward on the orchestration or tooling layers')
+    ->expect([
+        'Lattice\Lattice\Forms',
+        'Lattice\Lattice\Actions',
+        'Lattice\Lattice\Tables',
+        'Lattice\Lattice\Fragments',
+        'Lattice\Lattice\Layouts',
+    ])
+    ->not->toUse([
+        'Lattice\Lattice\Http',
+        'Lattice\Lattice\Pages',
+        'Lattice\Lattice\Console',
+    ]);
+
+/*
+ * Attributes are a shared base layer of plain markers: they describe domain
+ * objects without reaching into the domains. Actions is intentionally omitted
+ * while the effects system (AsEffect -> Actions\Enums\EffectType) is reworked.
+ */
+arch('attributes depend on no feature domain or higher layer')
+    ->expect('Lattice\Lattice\Attributes')
+    ->not->toUse([
+        'Lattice\Lattice\Forms',
+        'Lattice\Lattice\Tables',
+        'Lattice\Lattice\Fragments',
+        'Lattice\Lattice\Layouts',
+        'Lattice\Lattice\Http',
+        'Lattice\Lattice\Pages',
+        'Lattice\Lattice\Console',
+        'Lattice\Lattice\Facades',
+    ]);
+
 /*
  * Structural conventions.
  */
 
+/*
+ * Cross-boundary contracts live in a `Contracts` namespace and are interfaces.
+ * Capability interfaces that exist only to be implemented by a single local
+ * hierarchy (e.g. Tables\Columns\{Filterable, Sortable, ColumnProps},
+ * Support\TypeScript\TypeScriptProfile) deliberately sit beside their
+ * implementations and are not part of this convention.
+ */
 arch('contracts are interfaces')
     ->expect([
         'Lattice\Lattice\Core\Contracts',
